# internal/apigen/dexon_api.gen.go
// Package apigen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package apigen

import (
	"context"
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
)

const (
	ApiKeyAuthScopes = "ApiKeyAuth.Scopes"
)

// Defines values for VMStatus.
const (
	Deleting     VMStatus = "deleting"
	Failed       VMStatus = "failed"
	Pending      VMStatus = "pending"
	Provisioning VMStatus = "provisioning"
	Running      VMStatus = "running"
	Scheduled    VMStatus = "scheduled"
)

// CreateVMRequest defines model for CreateVMRequest.
type CreateVMRequest struct {
	// DiskImage Identifier of the base disk image to boot from
	DiskImage string `json:"disk_image"`

	// Flavor Name of a server-defined flavor (see GET /flavors)
	Flavor string `json:"flavor"`

	// Name Optional user-friendly label, not used for identity
	Name *string `json:"name,omitempty"`
}

// Flavor defines model for Flavor.
type Flavor struct {
	DiskGb   *int    `json:"disk_gb,omitempty"`
	MemoryMb *int    `json:"memory_mb,omitempty"`
	Name     *string `json:"name,omitempty"`
	Vcpus    *int    `json:"vcpus,omitempty"`
}

// VM defines model for VM.
type VM struct {
	CreatedAt *time.Time `json:"created_at,omitempty"`

	// FailureReason Present only when status is `failed`
	FailureReason *string    `json:"failure_reason,omitempty"`
	Flavor        *string    `json:"flavor,omitempty"`
	Id            *string    `json:"id,omitempty"`
	Name          *string    `json:"name,omitempty"`
	Status        *VMStatus  `json:"status,omitempty"`
	UpdatedAt     *time.Time `json:"updated_at,omitempty"`
}

// VMStatus defines model for VMStatus.
type VMStatus string

// ListVMsParams defines parameters for ListVMs.
type ListVMsParams struct {
	// Status Filter by lifecycle status
	Status *VMStatus `form:"status,omitempty" json:"status,omitempty"`
}

// CreateVMParams defines parameters for CreateVM.
type CreateVMParams struct {
	// Wait If true, hold the request open until the VM reaches a terminal state (running or failed) or timeout elapses. Defaults to false (return immediately).
	Wait *bool `form:"wait,omitempty" json:"wait,omitempty"`

	// Timeout Max time to wait when wait=true (e.g. "60s"). Ignored otherwise. If it elapses first, returns the current (still-pending) status rather than an error.
	Timeout *string `form:"timeout,omitempty" json:"timeout,omitempty"`
}

// CreateVMJSONRequestBody defines body for CreateVM for application/json ContentType.
type CreateVMJSONRequestBody = CreateVMRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// List available VM flavors
	// (GET /flavors)
	ListFlavors(w http.ResponseWriter, r *http.Request)
	// List VMs owned by the authenticated user
	// (GET /vms)
	ListVMs(w http.ResponseWriter, r *http.Request, params ListVMsParams)
	// Create a VM
	// (POST /vms)
	CreateVM(w http.ResponseWriter, r *http.Request, params CreateVMParams)
	// Destroy a VM
	// (DELETE /vms/{id})
	DeleteVM(w http.ResponseWriter, r *http.Request, id string)
	// Get VM status and details
	// (GET /vms/{id})
	GetVM(w http.ResponseWriter, r *http.Request, id string)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// List available VM flavors
// (GET /flavors)
func (_ Unimplemented) ListFlavors(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// List VMs owned by the authenticated user
// (GET /vms)
func (_ Unimplemented) ListVMs(w http.ResponseWriter, r *http.Request, params ListVMsParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Create a VM
// (POST /vms)
func (_ Unimplemented) CreateVM(w http.ResponseWriter, r *http.Request, params CreateVMParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Destroy a VM
// (DELETE /vms/{id})
func (_ Unimplemented) DeleteVM(w http.ResponseWriter, r *http.Request, id string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get VM status and details
// (GET /vms/{id})
func (_ Unimplemented) GetVM(w http.ResponseWriter, r *http.Request, id string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// ListFlavors operation middleware
func (siw *ServerInterfaceWrapper) ListFlavors(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	ctx = context.WithValue(ctx, ApiKeyAuthScopes, []string{})

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListFlavors(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// ListVMs operation middleware
func (siw *ServerInterfaceWrapper) ListVMs(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	ctx = context.WithValue(ctx, ApiKeyAuthScopes, []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params ListVMsParams

	// ------------- Optional query parameter "status" -------------

	err = runtime.BindQueryParameter("form", true, false, "status", r.URL.Query(), &params.Status)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "status", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListVMs(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// CreateVM operation middleware
func (siw *ServerInterfaceWrapper) CreateVM(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	ctx = context.WithValue(ctx, ApiKeyAuthScopes, []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params CreateVMParams

	// ------------- Optional query parameter "wait" -------------

	err = runtime.BindQueryParameter("form", true, false, "wait", r.URL.Query(), &params.Wait)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "wait", Err: err})
		return
	}

	// ------------- Optional query parameter "timeout" -------------

	err = runtime.BindQueryParameter("form", true, false, "timeout", r.URL.Query(), &params.Timeout)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "timeout", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateVM(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// DeleteVM operation middleware
func (siw *ServerInterfaceWrapper) DeleteVM(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithLocation("simple", false, "id", runtime.ParamLocationPath, chi.URLParam(r, "id"), &id)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	ctx = context.WithValue(ctx, ApiKeyAuthScopes, []string{})

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeleteVM(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// GetVM operation middleware
func (siw *ServerInterfaceWrapper) GetVM(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithLocation("simple", false, "id", runtime.ParamLocationPath, chi.URLParam(r, "id"), &id)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	ctx = context.WithValue(ctx, ApiKeyAuthScopes, []string{})

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetVM(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/flavors", wrapper.ListFlavors)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/vms", wrapper.ListVMs)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/vms", wrapper.CreateVM)
	})
	r.Group(func(r chi.Router) {
		r.Delete(options.BaseURL+"/vms/{id}", wrapper.DeleteVM)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/vms/{id}", wrapper.GetVM)
	})

	return r
}